pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speechdown"
version = "0.1.0"
description = "A CLI tool for transcribing audio files in multiple languages."
authors = [{ name = "Artem Dudarev", email = "dudarev@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "openai-whisper",
]

[project.scripts]
sd = "speechdown.main:cli"

[project.optional-dependencies]
testing = ["pytest", "ruff", "mypy"]

[tool.ruff]
line-length = 100
fix = true

[tool.setuptools]
packages = ["speechdown"]
package-dir = { "" = "src" }
